# docs/projects-service/openapi.yaml
openapi: 3.0.0
info:
  title: Projects Service API
  description: Microservice for managing user projects with collaborative filesystem structure
  version: 1.1.0
  contact:
    name: Projects Service Team
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /projects/v1/projects/{projectID}:
    get:
      summary: Get project details
      operationId: getProject
      tags:
        - Projects
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Update project name
      description: Update project name. Requires write permission.
      operationId: updateProject
      tags:
        - Projects
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: Updated Project Name
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete project
      description: Delete a project. Owner only.
      operationId: deleteProject
      tags:
        - Projects
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '204':
          description: Project deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /projects/v1/projects/{projectID}/directories/{dirID}/download:
    get:
      summary: Download a directory as a ZIP archive
      description: >
        Requires read access. The selected directory is the ZIP's top-level
        entry; nested and empty directories are preserved. Unsafe or ambiguous
        archive paths are rejected before streaming begins.
      operationId: downloadDirectory
      tags: [Downloads]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/DirID'
      responses:
        '200':
          $ref: '#/components/responses/ZipDownload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Unsafe, corrupt, duplicate, or ambiguous archive paths.
        '502':
          $ref: '#/components/responses/DownloadUnavailable'
  /projects/v1/projects/{projectID}/download:
    get:
      summary: Download a project as a ZIP archive
      description: >
        Requires read access. ZIP entries are relative to the project's root;
        no duplicate project-name directory is added.
      operationId: downloadProject
      tags: [Downloads]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '200':
          $ref: '#/components/responses/ZipDownload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Unsafe, corrupt, duplicate, or ambiguous archive paths.
        '502':
          $ref: '#/components/responses/DownloadUnavailable'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the authentication service
  parameters:
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    DirID:
      name: dirID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Project:
      type: object
      required:
        - id
        - owner_id
        - name
        - created_at
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          description: User ID of the project owner
        name:
          type: string
          maxLength: 255
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
  responses:
...
    ZipDownload:
      description: A private ZIP attachment with a safe RFC 6266 filename.
      headers:
        Cache-Control:
          schema: { type: string, example: "private, no-store" }
        Content-Disposition:
          schema: { type: string, example: "attachment; filename=project.zip; filename*=UTF-8''project.zip" }
      content:
        application/zip:
          schema: { type: string, format: binary }
    DownloadUnavailable:
      description: Storage or collaborative-text materialization could not be completed before the response started.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized — missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden — user lacks required permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...